+++ front-end/studio/src/app/models/api-template-kinds.enum.ts
