+++ front-end/studio/src/app/models/stored-api-design-template.model.ts
