+++ front-end/studio/src/app/models/stored-organisation.model.ts
@@ -1,4 +1,4 @@
-import { OrganisationModel } from "./organisation/organisation.model";
+import { OrganisationModel } from "./organisation.model";
 
 export interface StoredOrganisation extends OrganisationModel {
