+++ front-end/studio/src/app/models/template.model.ts
