+++ front-end/studio/src/app/pages/apis/_components/apis-list.component.ts
@@ -16,7 +16,7 @@
  */
 
 import {Component, EventEmitter, Input, Output} from "@angular/core";
-import {Api} from "../../../models/api/api.model";
+import {Api} from "../../../models/api.model";
 
 
 @Component({
