+++ front-end/studio/src/app/pages/apis/import/_components/import-form.component.ts
@@ -17,7 +17,7 @@
 
 import {Component, EventEmitter, Input, Output} from "@angular/core";
 import {ApisService} from "../../../../services/apis.service";
-import {ImportApi} from "../../../../models/api/import-api.model";
+import {ImportApi} from "../../../../models/import-api.model";
 import {DropDownOption, DropDownOptionValue as Value} from '../../../../components/common/drop-down.component';
 import {CodeEditorMode, CodeEditorTheme} from "../../../../components/common/code-editor.component";
 import {Base64} from "js-base64";
