+++ front-end/studio/src/app/pages/apis/{apiId}/collaboration/_components/invitation.component.ts
@@ -19,7 +19,7 @@ import {Component, QueryList, ViewChildren} from "@angular/core";
 import {ModalDirective} from "ngx-bootstrap/modal";
 import {Invitation} from "../../../../../models/invitation.model";
 import {ClipboardService} from "ngx-clipboard";
-import {Api} from "../../../../../models/api/api.model";
+import {Api} from "../../../../../models/api.model";
 
 
 @Component({
